refactor(aliases): split into commands/ + lib/ with a thin index.ts
