+++ src/aliases/lib/aliases.test.ts
